Moved ImageHelper.cs into a Utility folder and therefore into the PokeMemo.Utility namespace.
