+++ Models/Card.cs
@@ -1,10 +1,6 @@
 ﻿using Avalonia.Media.Imaging;
 using CommunityToolkit.Mvvm.ComponentModel;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using PokeMemo.Utility;
 
 namespace PokeMemo.Models
 {
