+++ Models/Deck.cs
@@ -1,6 +1,7 @@
 ﻿using CommunityToolkit.Mvvm.ComponentModel;
 using System.Collections.Generic;
 using Avalonia.Media.Imaging;
+using PokeMemo.Utility;
 
 namespace PokeMemo.Models
 {
